implementação singleton

+++ src/main/java/br/com/geradorprova/controller/QuestaoController.java
@@ -1,6 +1,7 @@
 package br.com.geradorprova.controller;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -15,7 +16,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 
 import br.com.geradorprova.model.Questao;
 import br.com.geradorprova.model.Resposta;
+import br.com.geradorprova.model.UserSingleton;
 import br.com.geradorprova.model.enumeration.TipoQuestao;
+import br.com.geradorprova.model.enumeration.TipoUsuario;
 import br.com.geradorprova.service.QuestaoService;
 
 @Controller
@@ -27,6 +30,11 @@ public class QuestaoController{
 
 	@GetMapping("/novo")
 	public String formNew(Model model) {
+		
+		if(Objects.isNull(UserSingleton.getInstance().getTipo()) || TipoUsuario.ALUNO.equals(UserSingleton.getInstance().getTipo())) {
+			return "redirect:/index";
+		}
+		
 		Questao questao = new Questao();
 //		questao.setRespostas(Stream.generate(Resposta::new).limit(5).collect(Collectors.toList()));
 		model.addAttribute(questao);
@@ -38,6 +46,10 @@ public class QuestaoController{
 	@PostMapping("/salvar")
 	public String save(Questao questao, Model model) {
 		
+		if(Objects.isNull(UserSingleton.getInstance().getTipo()) || TipoUsuario.ALUNO.equals(UserSingleton.getInstance().getTipo())) {
+			return "redirect:/index";
+		}
+		
 		questaoService.save(questao);
 		
 		return "redirect:/questao/listar";
@@ -46,6 +58,10 @@ public class QuestaoController{
 	@GetMapping("/deletar/{id}")
 	public String delete(@PathVariable Long id) {
 		
+		if(Objects.isNull(UserSingleton.getInstance().getTipo()) || TipoUsuario.ALUNO.equals(UserSingleton.getInstance().getTipo())) {
+			return "redirect:/index";
+		}
+		
 		questaoService.delete(id);
 		
 		return "redirect:/questao/listar";
@@ -53,6 +69,11 @@ public class QuestaoController{
 
 	@GetMapping("/editar/{id}")
 	public String formEdit(@PathVariable Long id, Model model) {
+		
+		if(Objects.isNull(UserSingleton.getInstance().getTipo()) || TipoUsuario.ALUNO.equals(UserSingleton.getInstance().getTipo())) {
+			return "redirect:/index";
+		}
+		
 		Questao questao = questaoService.findById(id);
 		
 		model.addAttribute("questao", questao);
@@ -63,6 +84,10 @@ public class QuestaoController{
 
 	@GetMapping("/listar")
 	public String list(Model model){
+		
+		if(Objects.isNull(UserSingleton.getInstance().getTipo())) {
+			return "redirect:/index";
+		}
 		
 		List<Questao> questoes = questaoService.listAll();
